LayoutTests/http/tests/websocket/tests/hybi/workers/close-code-and-reason-expected.txt

CONSOLE MESSAGE: WebSocket is closed before the connection is established.
Test CloseEvent code and reason property in Worker.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

ws.onclose() was called.
PASS PASS: worker: event.wasClean is false
PASS PASS: worker: event.code is codeAbnormalClosure
PASS PASS: worker: event.reason is emptyString
WebSocketTest.onopen() was called with testId = 0.
WebSocketTest.onclose() was called with testId = 0.
PASS PASS: worker: event.wasClean is true
PASS PASS: worker: event.code is 1005
PASS PASS: worker: event.reason is 
PASS PASS: worker: event.code is 1005
PASS PASS: worker: event.reason is 
WebSocketTest.onopen() was called with testId = 1.
WebSocketTest.onclose() was called with testId = 1.
PASS PASS: worker: event.wasClean is true
PASS PASS: worker: event.code is 1005
PASS PASS: worker: event.reason is 
PASS PASS: worker: event.code is 1005
PASS PASS: worker: event.reason is 
WebSocketTest.onopen() was called with testId = 2.
WebSocketTest.onclose() was called with testId = 2.
PASS PASS: worker: event.wasClean is true
PASS PASS: worker: event.code is 65535
PASS PASS: worker: event.reason is good bye
PASS PASS: worker: event.code is 65535
PASS PASS: worker: event.reason is good bye
DONE
PASS successfullyParsed is true

TEST COMPLETE
